+++ ci/checks/enforce-release-notes-up-to-date.sh
@@ -2,7 +2,8 @@
 
 set -e -o pipefail -u
 
-current_version=$(python3 setup.py --version)
+# Not using `python3 setup.py --version` to make it work in the contexts where setuptools are missing (Python 3.12+)
+current_version=$(cut -d\' -f2 git_machete/__init__.py)
 release_notes_version=$(sed '3!d' RELEASE_NOTES.md | grep -Eo '(0|[1-9][0-9]*)(\.(0|[1-9][0-9]*))+')
 
 if [[ $current_version != "$release_notes_version" ]]; then