+++ tests/test_github_retarget_pr.py
@@ -274,7 +274,6 @@ class TestGitHubRetargetPR(BaseTest):
             * PR #15: `feature` ➔ `master`
 
             <!-- end git-machete generated -->
-
             # Summary''')[1:]
 
         # branch feature_2 is not present in any of the remotes
@@ -321,8 +320,7 @@ class TestGitHubRetargetPR(BaseTest):
             * PR #25: `feature_2` ➔ `feature`
             * PR #15: `feature` ➔ `master`
 
-            <!-- end git-machete generated -->
-        ''')[1:]
+            <!-- end git-machete generated -->''')[1:]
 
         # branch feature_3 present in only one remote: origin_1 and has tracking data
         (
@@ -364,7 +362,6 @@ class TestGitHubRetargetPR(BaseTest):
             * PR #15: `feature` ➔ `master`
 
             <!-- end git-machete generated -->
-
             # Summary''')[1:]
 
         body = \
@@ -398,7 +395,6 @@ class TestGitHubRetargetPR(BaseTest):
             * PR #15: `feature` ➔ `master`
 
             <!-- end git-machete generated -->
-
             # Summary''')[1:]
 
         body = \